review#8 : separate functions to module
