+++ index.html
@@ -34,7 +34,7 @@
       </div>
     </main>
   </div>
-  <script src="js/index.js"></script>
+  <script src="js/index.js" type="module"></script>
 
 </body>
